# openapi.yaml
openapi: 3.0.3
info:
  title: MoWizz API
  version: 1.0.0
  description: OpenAPI documentation for current endpoints.
servers:
  - url: http://localhost:8080
paths:
  /movie/search:
    get:
      summary: Search movies by query
      parameters:
        - in: query
          name: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Matching movies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TmdbMovie"
  /movie/{tmdbId}:
    get:
      summary: Get a TMDB movie by TMDB id
      parameters:
        - in: path
          name: tmdbId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Movie found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TmdbMovie"
...
components:
  schemas:
...
    TmdbMovie:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        release_date:
          type: string
          format: date
        poster_path:
          type: string
        overview:
          type: string
        vote_average:
          type: number
          format: double