# src/main/java/com/github/davidduclam/movietracker/controller/MovieController.java
package com.github.davidduclam.movietracker.controller;

import com.github.davidduclam.movietracker.dto.TmdbMovieDTO;
import com.github.davidduclam.movietracker.model.Movie;
import com.github.davidduclam.movietracker.service.MovieService;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping("/movie")
public class MovieController {
    private final MovieService movieService;

    public MovieController(MovieService movieService) {
        this.movieService = movieService;
    }

    @GetMapping("/search")
    public List<TmdbMovieDTO> search(@RequestParam String query) {
        return movieService.searchMovie(query);
    }

//    @GetMapping("/{tmdbId}")
//    public Movie saveMovie(@PathVariable Long tmdbId) {
//        return movieService.getOrCreateMovie(tmdbId);
//    }

    @GetMapping("/{tmdbId}")
    public TmdbMovieDTO findMovie(@PathVariable Long tmdbId) {
        return movieService.findMovie(tmdbId);
    }

    @GetMapping("/popular")
    public List<TmdbMovieDTO> popularMovies() {
        return movieService.popularMovies();
    }

    @GetMapping("/top-rated")
    public List<TmdbMovieDTO> topRatedMovies() {
        return movieService.topRatedMovies();
    }

    @GetMapping("/upcoming")
    public List<TmdbMovieDTO> upcomingMovies() {
        return movieService.upcomingMovies();
    }
}